# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Genshin API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/artifacts": {
            "get": {
                "tags": [
                    "Artifacts"
                ],
                "summary": "Get characters based on parameters",
                "description": "Get characters based on specified parameters.",
                "operationId": "80643be7a3a6fe5c71851277acfd7cf4",
                "parameters": [
                    {
                        "name": "stat",
                        "in": "query",
                        "description": "The stat of the set artifacts.",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of artifacts filtered.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "List of artifacts filtered."
                                        },
                                        "artifacts": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No artifacts found with the provided parameters.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No artifacts found with the provided parameters."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/artifact/{name}": {
            "get": {
                "tags": [
                    "Artifacts"
                ],
                "summary": "Get an artifact by name",
                "description": "Get a artifact by name.",
                "operationId": "46a54d2f28caab164f4deb9bb02da31f",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "The name of the artifact.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Artifact found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Artifact found."
                                        },
                                        "artifact": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No artifact found with the provided name.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No artifact found with the provided name."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/characters": {
            "get": {
                "tags": [
                    "Characters"
                ],
                "summary": "Get characters based on parameters",
                "description": "Get characters based on specified parameters.",
                "operationId": "fe9645e9336066f4ac20d5109b2bac7b",
                "parameters": [
                    {
                        "name": "area",
                        "in": "query",
                        "description": "The area of the character.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "rarity",
                        "in": "query",
                        "description": "The rarity of the character.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "weapon",
                        "in": "query",
                        "description": "The weapon type of the character.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "element",
                        "in": "query",
                        "description": "The element of the character.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ascension",
                        "in": "query",
                        "description": "The ascension stat of the character.",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of characters filtered.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Liste de personnages filtrée."
                                        },
                                        "characters": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No characters found with the provided parameters.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No characters found with the provided parameters."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/character/{name}": {
            "get": {
                "tags": [
                    "Characters"
                ],
                "summary": "Get a character by name",
                "description": "Get a character by name.",
                "operationId": "ae162bed5089267ec653a55354d2fd0d",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "The name of the character.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Character found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Character found."
                                        },
                                        "character": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No character found with the provided name.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No character found with the provided name."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/weapons": {
            "get": {
                "tags": [
                    "Weapons"
                ],
                "summary": "Get weapons based on parameters",
                "description": "Get characters based on specified parameters.",
                "operationId": "ab21959c64abd6925ea0fee48342bbb6",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "The type of the weapon.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "rarity",
                        "in": "query",
                        "description": "The rarity of the weapon.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ascension",
                        "in": "query",
                        "description": "The ascension of the weapon.",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of weapons filtered.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "List of weapons filtered."
                                        },
                                        "weapons": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No weapons found with the provided parameters.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No weapons found with the provided parameters."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/weapon/{name}": {
            "get": {
                "tags": [
                    "Weapons"
                ],
                "summary": "Get a weapon by name",
                "description": "Get a weapon by name.",
                "operationId": "c58e08b303cc1202078413f37fa0cf68",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "The name of the weapon.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Weapon found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Weapon found."
                                        },
                                        "weapon": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No weapon found with the provided name.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No weapon found with the provided name."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}